+++ resources/views/Edit.blade.php
@@ -9,6 +9,7 @@
     <link rel="stylesheet"
         href="{{ asset('css/autocomplete.css') }}?v={{ filemtime(public_path('css/autocomplete.css')) }}">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/7.0.1/css/all.min.css">
+    <x-title-tooltip-assets />
 </head>
 
 <body class="product-form-page">