+++ resources/views/Options.blade.php
@@ -12,6 +12,7 @@
     <link rel="stylesheet"
         href="{{ asset('css/list-menu-float.css') }}?v={{ filemtime(public_path('css/list-menu-float.css')) }}">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/7.0.1/css/all.min.css">
+    <x-title-tooltip-assets />
 
     @livewireStyles
 </head>