+++ resources/views/TagLibrary.blade.php
@@ -15,6 +15,7 @@
     <link rel="stylesheet"
         href="{{ asset('css/list-menu-float.css') }}?v={{ filemtime(public_path('css/list-menu-float.css')) }}">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/7.0.1/css/all.min.css">
+    <x-title-tooltip-assets />
 
     @livewireStyles
 </head>
@@ -37,7 +38,8 @@
     </div>
 
     @livewireScripts
-    <script src="{{ asset('scripts/autocomplete-text.js') }}?v={{ filemtime(public_path('scripts/autocomplete-text.js')) }}">
+    <script
+        src="{{ asset('scripts/autocomplete-text.js') }}?v={{ filemtime(public_path('scripts/autocomplete-text.js')) }}">
     </script>
 </body>