+++ resources/views/components/title-tooltip-assets.blade.php
@@ -1,0 +1,4 @@
+<link rel="stylesheet"
+    href="{{ asset('css/title-tooltips.css') }}?v={{ filemtime(public_path('css/title-tooltips.css')) }}">
+<script src="{{ asset('scripts/title-tooltips.js') }}?v={{ filemtime(public_path('scripts/title-tooltips.js')) }}"
+    defer></script>